samples/http-image-cloudevents/requirements.txt
flask
requests
Pillow
pytest
base64